committing to not lose progress
